app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "debugpy>=1.8.8",
    "fastapi[all]>=0.115.4",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.0.1",
    "redis-om>=0.3.3",
    "redis>=5.2.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "uvicorn>=0.32.0",
]
